Update gperf to v3.2
The gperf 3.2 release happened early this year but the release wasn't packaged on their ftp. Use a GitHub mirror of the sources instead.

Build is modified to use a static runtime on Windows. It also targets Windows 8.1. This is to make sure that the executable runs within the docker-webkit-dev containers.